Create gmaps-block.js
Currently a work in progress, I can currently render the map in a text box but can't get it to pull back and render with it's own pastable part. 

Is mostly based off the Sir Trev video block since thats got the draggable/pastable features already set. Just need to get the regex to process the map correctly now.